Audit fixes: sync integrity, int-cents money, AI via backend, secure entitlement
Sync:
- Mark only server-accepted records synced (stop silent data loss); real sync mutex
- Fix pull field mappings (IconName / MainCategoryId); guard each pulled record
- Use server ServerTime as the pull cursor
- Transaction delete -> soft-delete (syncs); wallet delete cascades to its transactions

Money -> integer minor units (cents):
- All money double -> int; Drift schema v11 migration (x100) + openingBalance backfill
- calculateWalletBalance rule aligned with migration (isPaid OR credit/debt)
- Regenerated Drift code; removed stale analyzer dependency override

Credit/debt + edit:
- Outstanding tracked by paidAmount (not the overloaded isPaid flag)
- Transaction-edit balance predicate includes credit/debt

Premium/AI:
- Route premium AI through backend (/ai/insights); drop bundled Gemini key
- Move cached premium entitlement to flutter_secure_storage (migrate off plaintext)

+++ lib/features/transactions/providers/transfer_provider.dart
@@ -70,7 +70,7 @@ class TransferNotifier extends StateNotifier<TransferState> {
   Future<bool> createTransfer({
     required String sourceWalletId,
     required String destinationWalletId,
-    required double amount,
+    required int amount, // integer minor units / cents
     required DateTime date,
     String? notes,
     String? title,
@@ -120,7 +120,7 @@ class TransferNotifier extends StateNotifier<TransferState> {
   /// Update an existing transfer
   Future<bool> updateTransfer({
     required String transactionId,
-    double? amount,
+    int? amount, // integer minor units / cents
     DateTime? date,
     String? notes,
     String? title,
